Create Delete GET

+++ webapp/students/routes.py
@@ -190,40 +190,44 @@ def update_student():
 
 
 
-@student.route('/delete-student', methods=['POST'])
+@student.route('/delete-student', methods=['GET', 'POST'])
 def deleteRecord():
-    data = request.get_json()
-    print(data) 
-    student_name = data['studentName']
-    subject_name = data.get('subject', None)
-    print(student_name, subject_name)
+    if request.method == 'POST':
+        data = request.get_json()
+        print(data) 
+        student_name = data['studentName']
+        subject_name = data.get('subject', None)
+        print(student_name, subject_name)
 
-    # Student should exist - to delete
-    existing_student = db.mst_student.find_one({"student_name": student_name})
-    if not existing_student:
-        # if error occured send a error message
-        return {'message': 'Student with name '+student_name+' does not exist!',  'category':'danger'}
+        # Student should exist - to delete
+        existing_student = db.mst_student.find_one({"student_name": student_name})
+        if not existing_student:
+            # if error occured send a error message
+            return {'message': 'Student with name '+student_name+' does not exist!',  'category':'danger'}
 
-     # If only student name is provided, delete all occurrences with that name
-    if not subject_name:
-        result = db.mst_student.delete_many({"student_name": student_name})
-        if result.deleted_count > 0:
-            return {'message': 'Sucess, Deleted all records of '+student_name+'!',  'category':'success'}
-        else:
-            return {'message': 'Error occured',  'category':'danger'}
-    else:
-        # Check if the subject exists in the subject table
-        subject_collection = db.mst_subject
-        existing_subject = subject_collection.find_one({"subject_name": subject_name})
-
-        if existing_subject:
-            subject_id = existing_subject['_id']
-            # If both student name and subject are provided, delete the specific record
-            result = db.mst_student.delete_one({"student_name": student_name, "subject_key": subject_id})
+        # If only student name is provided, delete all occurrences with that name
+        if not subject_name:
+            result = db.mst_student.delete_many({"student_name": student_name})
             if result.deleted_count > 0:
-                return {'message': 'Sucess, Deleted a specific record of '+student_name+'!',  'category':'success'}
+                return {'message': 'Sucess, Deleted all records of '+student_name+'!',  'category':'success'}
+            else:
+                return {'message': 'Error occured',  'category':'danger'}
         else:
-            return {'message': 'Subject name provided does not exist (Deleting records is case sensitive)',  'category':'danger'}
+            # Check if the subject exists in the subject table
+            subject_collection = db.mst_subject
+            existing_subject = subject_collection.find_one({"subject_name": subject_name})
+
+            if existing_subject:
+                subject_id = existing_subject['_id']
+                # If both student name and subject are provided, delete the specific record
+                result = db.mst_student.delete_one({"student_name": student_name, "subject_key": subject_id})
+                if result.deleted_count > 0:
+                    return {'message': 'Sucess, Deleted a specific record of '+student_name+'!',  'category':'success'}
+            else:
+                return {'message': 'Subject name provided does not exist (Deleting records is case sensitive)',  'category':'danger'}
+    elif request.method == 'GET':
+        # Handle GET request here
+        return render_template('update.html')
 
 
 def addSubjectName(studs):
@@ -235,6 +239,7 @@ def addSubjectName(studs):
         if subject:
             student['subject_name'] = subject['subject_name']
             student.pop('_id', None)
+            student.pop('subject_key', None)
             print("Added sub name", student)
         else:
             print("no subject= ", student['subject_key'])
